Move output file generation to separate plugin
